ai_writer/requirements.txt
openai-agents==0.0.16
mysql-connector-python
python-dotenv
requests 